+++ src/scenes.js
@@ -51,7 +51,7 @@
 /*global modules, VariableFrame, StageMorph, SpriteMorph, Process, List,
 normalizeCanvas, SnapSerializer*/
 
-modules.scenes = '2021-July-02';
+modules.scenes = '2021-July-03';
 
 
 // Projecct /////////////////////////////////////////////////////////
@@ -180,7 +180,6 @@ Scene.prototype.addDefaultSprite = function () {
 
 Scene.prototype.captureGlobalSettings = function () {
     this.hiddenPrimitives = StageMorph.prototype.hiddenPrimitives;
-    this.unifiedPalette = StageMorph.prototype.unifiedPalette;
     this.codeMappings = StageMorph.prototype.codeMappings;
     this.codeHeaders = StageMorph.prototype.codeHeaders;
     this.enableCodeMapping = StageMorph.prototype.enableCodeMapping;
@@ -194,7 +193,6 @@ Scene.prototype.captureGlobalSettings = function () {
 
 Scene.prototype.applyGlobalSettings = function () {
     StageMorph.prototype.hiddenPrimitives = this.hiddenPrimitives;
-    StageMorph.prototype.unifiedPalette = this.unifiedPalette;
     StageMorph.prototype.codeMappings = this.codeMappings;
     StageMorph.prototype.codeHeaders = this.codeHeaders;
     StageMorph.prototype.enableCodeMapping = this.enableCodeMapping;