+++ HISTORY.md
@@ -16,7 +16,8 @@
 
 ### 2021-07-03
 * objects: arranged the blocks in the unified palette column-wise
-* gui: fixed the paletteHandle for the unified palette 
+* gui: fixed the paletteHandle for the unified palette
+* gui, scenes, store: fixed unified palette for multi-scene projects
 
 ### 2021-07-02
 * gui, object, store, etc.: unified blocks palette option, thanks, Michael!